+++ src/ts/ui/initUI.ts
@@ -26,9 +26,9 @@ export const initUI = (vditor: IVditor) => {
     if (vditor.toolbar.elements.outline) {
         const outlineElement = document.createElement("div");
         outlineElement.className = "vditor-outline";
-        let top = 0
+        let top = 0;
         if (vditor.options.toolbarConfig.pin) {
-            top = vditor.toolbar.element.clientHeight
+            top = vditor.toolbar.element.clientHeight;
         }
         outlineElement.innerHTML = `<div style='top:${top}px'></div>`;
         contentElement.appendChild(outlineElement);